requirements.txt
MyQR==2.0.0
pytest
requests
PyYAML